Use "racheting" search instead of binary search
Basically, you sort both of the lists, then keep track of your location
in both lists. Then, you loop, increasing only one of the indices, the
one where the value is lower. If values at both indices match, it is
added to the list of duplicates.

Surprisingly, this results in another speed up, going from 0.05s to
0.02s. I couldn't say why, except as a wild guess that cache coherency
is even better than with a binary search ¯\_(ツ)_/¯

It's also possible that the `list.sort()` method is faster than doing
a binary search in python.